backend/requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
gpxpy==1.6.2
requests==2.31.0
numpy==1.26.2
python-dotenv==1.0.0
google-generativeai==0.3.2
